requirements.txt
# Core dependencies
boto3>=1.26.0
jinja2>=3.0.0
pyyaml>=6.0
jsonschema>=4.0.0

# NLP and AI
spacy>=3.5.0
transformers>=4.26.0
nltk>=3.7

# Development and testing
pytest>=7.0.0
black>=23.1.0
flake8>=6.0.0
pytest-cov>=4.0.0

# Utilities
psutil>=5.9.0
requests>=2.28.0

# System monitoring and management
docker>=6.0.0

# Code quality analysis
bandit>=1.7.5
pylint>=2.16.0

# AWS
awscli>=1.27.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0